+++ CHANGELOG.md
@@ -93,6 +93,7 @@
 * Disable View source button for Source = FOLIO instances. Fixes UIIN-2570.
 * Add Inventory Browse facet for Call numbers (item) on Shared vs Not shared Instances. Refs UIIN-2407.
 * Add Browse facet for members with holdings on Instances in Contributors & Subject Browse. Refs UIIN-2415.
+* Add `Held By` facet in Call number browse. Refs UIIN-2416.
 
 ## [9.4.11](https://github.com/folio-org/ui-inventory/tree/v9.4.11) (2023-08-02)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v9.4.10...v9.4.11)